requirements.txt
black==21.9b0
click==8.0.3
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
regex==2021.10.23
tomli==1.2.1
typing-extensions==3.10.0.2
Werkzeug==2.0.2
investpy==1.0.8
python-binance==1.0.16
